+++ apps/api/src/app/account/account.controller.ts
@@ -152,7 +152,7 @@ export class AccountController {
       return this.accountService.createAccount(
         {
           ...data,
-          Platform: { connect: { id: platformId } },
+          platform: { connect: { id: platformId } },
           user: { connect: { id: this.request.user.id } }
         },
         this.request.user.id
@@ -250,7 +250,7 @@ export class AccountController {
         {
           data: {
             ...data,
-            Platform: { connect: { id: platformId } },
+            platform: { connect: { id: platformId } },
             user: { connect: { id: this.request.user.id } }
           },
           where: {
@@ -270,7 +270,7 @@ export class AccountController {
         {
           data: {
             ...data,
-            Platform: originalAccount.platformId
+            platform: originalAccount.platformId
               ? { disconnect: true }
               : undefined,
             user: { connect: { id: this.request.user.id } }