+++ apps/api/src/app/order/order.service.ts
@@ -490,7 +490,7 @@ export class OrderService {
           // eslint-disable-next-line @typescript-eslint/naming-convention
           Account: {
             include: {
-              Platform: true
+              platform: true
             }
           },
           // eslint-disable-next-line @typescript-eslint/naming-convention