+++ libs/common/src/lib/types/account-with-value.type.ts
@@ -2,7 +2,7 @@ import { Account as AccountModel, Platform } from '@prisma/client';
 
 export type AccountWithValue = AccountModel & {
   balanceInBaseCurrency: number;
-  Platform?: Platform;
+  platform?: Platform;
   transactionCount: number;
   value: number;
   valueInBaseCurrency: number;